+++ Locale/deDE.lua
@@ -1,28 +1,35 @@
 local L = LibStub("AceLocale-3.0"):NewLocale("SortGroup", "deDE")
-if not L then return end
+if not L then
+    return
+end
 
 L["SortGroup_Main_cb_Top_Text"] = "Spieler oben"
 L["SortGroup_Main_cb_Bottom_Text"] = "Spieler unten"
 L["SortGroup_Main_cb_Middle_Text"] = "Spieler in der Mitte"
 L["SortGroup_Main_cb_Descending_Text"] = "Absteigend"
 L["SortGroup_Main_cb_Ascending_Text"] = "Aufsteigend"
-L["SortGroup_Main_cb_Party1Top_Text"] = "Party1 oben"
-L["SortGroup_Main_cb_Party2Top_Text"] = "Party2 oben"
+L["SortGroup_Main_cb_UnevenTop_Text"] = "Ungerade oben"
+L["SortGroup_Main_cb_UnevenBottom_Text"] = "Ungerade unten"
+L["SortGroup_SortBy_Text"] = "Sortieren mit"
+L["SortGroup_Option_Text"] = "Optionen"
 L["SortGroup_Option_cb_ChatMessagesOn_Text"] = "Chat Nachrichten"
+L["SortGroup_Option_cb_HideInactiveSlots_Text"] = "Verstekce inaktive Gruppenmitglieder"
 L["SortGroup_Main_cb_Top_ToolTip"] = "Du bist an erster Stelle deiner Gruppe"
 L["SortGroup_Main_cb_Bottom_ToolTip"] = "Du bist an letzter Stelle deiner Gruppe"
-L["SortGroup_Main_cb_Middle_ToolTip"] = "[Experimentel]\nDu bist in der Mitte deiner Gruppe\n\nHinweis:\nDiese Option macht nur in Arena Sinn"
+L["SortGroup_Main_cb_Middle_ToolTip"] = "Du bist in der Mitte deiner Gruppe"
 L["SortGroup_Main_cb_Descending_ToolTip"] = "Party 1-4 werden nach unten sortiert"
 L["SortGroup_Main_cb_Ascending_ToolTip"] = "Party 1-4 werden nach oben sortiert"
-L["SortGroup_Main_cb_Party1Top_ToolTip"] = "Party1 an erster Stelle und Party2 an letzter Stelle deiner Gruppe"
-L["SortGroup_Main_cb_Party2Top_ToolTip"] = "Party2 an erster Stelle und Party1 an letzter Stelle deiner Gruppe"
+L["SortGroup_Main_cb_UnevenTop_ToolTip"] = "Party1 an erster Stelle und Party2 an letzter Stelle deiner Gruppe"
+L["SortGroup_Main_cb_UnivenBottom_ToolTip"] = "Party2 an erster Stelle und Party1 an letzter Stelle deiner Gruppe"
 L["SortGroup_Main_cb_ChatMessagesOn_ToolTip"] = "Chat Nachrichten an/aus";
+L["SortGroup_Option_cb_HideInactiveSlots_Tooltip"] =
+    "Wenn aktiv, alle nicht existierenden Gruppenmitglieder werden ausgeblendet"
 L["SortGroup_sort_top_descending_output"] = "SortGroup: Aktive mit - Oben, Absteigend"
 L["SortGroup_sort_top_ascending_output"] = "SortGroup: Aktive mit - Oben, Aufsteigend"
 L["SortGroup_sort_bottom_descending_output"] = "ortGroup: Aktive mit - Unten, Absteigend"
 L["SortGroup_sort_bottom_ascending_output"] = "SortGroup: \"'replacement'\" sortiert mit - Unten, Aufsteigend"
-L["SortGroup_sort_top_party1_output"] = "SortGroup: Aktive mit - Party1, Oben"
-L["SortGroup_sort_top_party2_output"] = "SortGroup: Aktive mit - Party2, Oben"
+L["SortGroup_sort_uneven_top_output"] = "SortGroup: Aktive mit - Party1, Oben"
+L["SortGroup_sort_uneven_bottom_output"] = "SortGroup: Aktive mit - Party2, Oben"
 L["SortGroup_sort_chat_Messages_On_output"] = "SortGroup: Chat Nachrichten sind an"
 L["SortGroup_sort_no_output"] = "SortGroup: Sortierfunktion ist nicht aktiv"
 L["SortGroup_in_combat_options_output"] = "SortGroup: Verlasse den Kampf, um Optionen ändern zu können"