+++ changelog.md
@@ -1,3 +1,10 @@
+5.2.0
+	Complete rework of the way the sorting gets applied
+		'filter_DB' can be changed to make unique cases
+	Added the player in the middle option
+	Added an option show/hide inactive player slots
+	Some code improvements
+
 5.1.07
 
 	Fixed a bug which caused the options to not load properly